修复Go编译错误和Docker构建问题

+++ server/api/v1/user/recharge.go
@@ -76,7 +76,7 @@ func CreateRechargeOrder(c *gin.Context) {
 	order := orderModel.Order{
 		OrderNo:       orderNo,
 		UserID:        userID.(uint),
-		Amount:        params.Amount,
+		Amount:        float64(params.Amount),
 		Status:        orderModel.OrderStatusPending,
 		PaymentMethod: params.PaymentMethod,
 		PaidAmount:    0,
